fix simulator issues found by the audit sweep
A recognized container that cannot be decoded is no longer replayed as
raw records: the probes swallow only what means "not this format", so a
corrupt xz header and a valid 7z fail instead of becoming fabricated
events. The rewriter refuses to write onto one of its inputs, which it
truncated before the lazily opened reader had read a byte, and both
CLIs exit with Picocli's status rather than reporting success for a
rejected command.

The combined report's rows are the union of its inputs, so a policy
reported at one size and not another is no longer dropped from the
chart, and the clairvoyant oracle rejects an event carrying both a
weight and penalties rather than materializing it without the
penalties.

TinyLfu rejects the weighted zero seed when its sketch cannot grow out
of it, which left count-min-64 aging its table on every increment. A
maximum size of zero is rejected once in the registry, where it was
accepted and not implemented. Three counters that wrap are widened to
long, the capacity scaling multiplies exactly so its existing
toIntExact sees the real value, and AccessEvent compares its penalties
the way it hashes them.

Note why the clock-pro policies retain an entry after its node leaves
the clock, since dropping it diverges from the reference's hit rate.

+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/policy/greedy_dual/GdsfPolicy.java
@@ -100,7 +100,7 @@ public final class GdsfPolicy implements Policy {
     priorityQueue.add(node);
   }
 
-  private double priorityOf(AccessEvent event, int frequency) {
+  private double priorityOf(AccessEvent event, long frequency) {
     // As defined in "Improving Web Servers and Proxies Performance with GDSF Caching Policies"
     double cost = event.isPenaltyAware() ? event.missPenalty() : 1.0;
     return clock + frequency * (cost / event.weight());
@@ -189,7 +189,7 @@ public final class GdsfPolicy implements Policy {
     final long key;
 
     double priority;
-    int frequency;
+    long frequency;
     int weight;
 
     public Node(long key, int weight, double priority) {
